pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ha-terra-listens"
version = "0.1.0"
description = "Home Assistant integration for Terra Listens bird monitoring"
requires-python = ">=3.12"
dependencies = [
    "terra-sdk @ git+https://github.com/stgarrity/terra-api.git@master",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-homeassistant-custom-component>=0.13",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
